moved all decomposition related method to test_decompositions.py
